using floats instead of ints for movement

+++ src/types.rs
@@ -2,24 +2,24 @@
 pub struct Entity {
     pub(crate) x: i32,
     pub(crate) y: i32,
-    pub(crate) vel_x: i32,
-    pub(crate) vel_y: i32,
-    pub(crate) mv_x: i32,
-    pub(crate) mv_y: i32,
-    pub(crate) sp_x: i32,
-    pub(crate) sp_y: i32,
+    pub(crate) vel_x: f32,
+    pub(crate) vel_y: f32,
+    pub(crate) mv_x: f32,
+    pub(crate) mv_y: f32,
+    pub(crate) sp_x: f32,
+    pub(crate) sp_y: f32,
 }
 
 impl Entity {
     pub fn new(
         x: i32,
         y: i32,
-        vel_x: i32,
-        vel_y: i32,
-        mv_x: i32,
-        mv_y: i32,
-        sp_x: i32,
-        sp_y: i32,
+        vel_x: f32,
+        vel_y: f32,
+        mv_x: f32,
+        mv_y: f32,
+        sp_x: f32,
+        sp_y: f32,
     ) -> Self {
         Entity {
             x,
